+++ dst/jobs/continuous/refundAccount/run.js
@@ -13,16 +13,17 @@ Object.defineProperty(exports, "__esModule", { value: true });
  */
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
+const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../../mongooseConnectionOptions");
 const debug = createDebug('sskts-jobs:*');
-sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default)
+mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default)
     .then(debug)
     // tslint:disable-next-line:no-console
     .catch(console.error);
 let count = 0;
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 200;
-const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
+const taskRepo = new sskts.repository.Task(mongoose.connection);
 const authClient = new sskts.pecorinoapi.auth.ClientCredentials({
     domain: process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
     clientId: process.env.PECORINO_API_CLIENT_ID,
@@ -38,7 +39,7 @@ setInterval(() => __awaiter(this, void 0, void 0, function* () {
     try {
         yield sskts.service.task.executeByName(sskts.factory.taskName.RefundAccount)({
             taskRepo: taskRepo,
-            connection: sskts.mongoose.connection,
+            connection: mongoose.connection,
             pecorinoEndpoint: process.env.PECORINO_API_ENDPOINT,
             pecorinoAuthClient: authClient
         });