+++ dst/jobs/continuous/returnPointAward/run.js
@@ -9,29 +9,38 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 /**
- * ムビチケ着券オーソリキャンセル
+ * Pecorinoインセンティブ返却
  */
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
+const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../../mongooseConnectionOptions");
 const debug = createDebug('sskts-jobs:*');
-sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default)
+mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default)
     .then(debug)
     // tslint:disable-next-line:no-console
     .catch(console.error);
 let count = 0;
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 1000;
-const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
+const INTERVAL_MILLISECONDS = 200;
+const taskRepo = new sskts.repository.Task(mongoose.connection);
+const authClient = new sskts.pecorinoapi.auth.ClientCredentials({
+    domain: process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
+    clientId: process.env.PECORINO_API_CLIENT_ID,
+    clientSecret: process.env.PECORINO_API_CLIENT_SECRET,
+    scopes: [],
+    state: ''
+});
 setInterval(() => __awaiter(this, void 0, void 0, function* () {
     if (count > MAX_NUBMER_OF_PARALLEL_TASKS) {
         return;
     }
     count += 1;
     try {
-        yield sskts.service.task.executeByName(sskts.factory.taskName.CancelMvtk)({
+        yield sskts.service.task.executeByName(sskts.factory.taskName.ReturnPointAward)({
             taskRepo: taskRepo,
-            connection: sskts.mongoose.connection
+            connection: mongoose.connection,
+            pecorinoAuthClient: authClient
         });
     }
     catch (error) {