+++ dst/jobs/triggered/importMovieTheaters/run.js
@@ -13,32 +13,35 @@ Object.defineProperty(exports, "__esModule", { value: true });
  */
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
+const mongoose = require("mongoose");
 const mongooseConnectionOptions_1 = require("../../../mongooseConnectionOptions");
 const debug = createDebug('sskts-jobs:jobs');
 function main() {
     return __awaiter(this, void 0, void 0, function* () {
         debug('connecting mongodb...');
-        yield sskts.mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
-        const organizationRepo = new sskts.repository.Organization(sskts.mongoose.connection);
-        const placeRepo = new sskts.repository.Place(sskts.mongoose.connection);
+        yield mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions_1.default);
+        const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+        const placeRepo = new sskts.repository.Place(mongoose.connection);
         // 全劇場組織を取得
-        const movieTheaters = yield organizationRepo.searchMovieTheaters({});
-        for (const movieTheater of movieTheaters) {
-            const branchCode = movieTheater.location.branchCode;
-            try {
-                debug('importing movieTheater...', branchCode);
-                yield sskts.service.masterSync.importMovieTheater(branchCode)({
-                    organization: organizationRepo,
-                    place: placeRepo
-                });
-                debug('movieTheater imported', branchCode);
-            }
-            catch (error) {
-                // tslint:disable-next-line:no-console
-                console.error(error);
+        const sellers = yield sellerRepo.search({});
+        for (const seller of sellers) {
+            if (seller.location !== undefined && seller.location.branchCode !== undefined) {
+                try {
+                    const branchCode = seller.location.branchCode;
+                    debug('importing movieTheater...', branchCode);
+                    yield sskts.service.masterSync.importMovieTheater(branchCode)({
+                        seller: sellerRepo,
+                        place: placeRepo
+                    });
+                    debug('movieTheater imported', branchCode);
+                }
+                catch (error) {
+                    // tslint:disable-next-line:no-console
+                    console.error(error);
+                }
             }
         }
-        yield sskts.mongoose.disconnect();
+        yield mongoose.disconnect();
     });
 }
 main()