+++ src/jobs/continuous/cancelPointAward/run.ts
@@ -1,14 +1,15 @@
 /**
- * ムビチケ着券オーソリキャンセル
+ * Pecorinoインセンティブ承認取消
  */
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../mongooseConnectionOptions';
 
 const debug = createDebug('sskts-jobs:*');
 
-sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions)
+mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions)
     .then(debug)
     // tslint:disable-next-line:no-console
     .catch(console.error);
@@ -16,8 +17,16 @@ sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptio
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 1000;
-const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
+const INTERVAL_MILLISECONDS = 200;
+const taskRepo = new sskts.repository.Task(mongoose.connection);
+
+const authClient = new sskts.pecorinoapi.auth.ClientCredentials({
+    domain: <string>process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
+    clientId: <string>process.env.PECORINO_API_CLIENT_ID,
+    clientSecret: <string>process.env.PECORINO_API_CLIENT_SECRET,
+    scopes: [],
+    state: ''
+});
 
 setInterval(
     async () => {
@@ -29,10 +38,11 @@ setInterval(
 
         try {
             await sskts.service.task.executeByName(
-                sskts.factory.taskName.CancelMvtk
+                sskts.factory.taskName.CancelPointAward
             )({
                 taskRepo: taskRepo,
-                connection: sskts.mongoose.connection
+                connection: mongoose.connection,
+                pecorinoAuthClient: authClient
             });
         } catch (error) {
             // tslint:disable-next-line:no-console