+++ src/jobs/continuous/returnPointAward/run.ts
@@ -1,15 +1,15 @@
 /**
- * ムビチケ使用タスク監視
- * 実際は何もしない
+ * Pecorinoインセンティブ返却
  */
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../mongooseConnectionOptions';
 
 const debug = createDebug('sskts-jobs:*');
 
-sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions)
+mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions)
     .then(debug)
     // tslint:disable-next-line:no-console
     .catch(console.error);
@@ -18,7 +18,15 @@ let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 200;
-const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
+const taskRepo = new sskts.repository.Task(mongoose.connection);
+
+const authClient = new sskts.pecorinoapi.auth.ClientCredentials({
+    domain: <string>process.env.PECORINO_AUTHORIZE_SERVER_DOMAIN,
+    clientId: <string>process.env.PECORINO_API_CLIENT_ID,
+    clientSecret: <string>process.env.PECORINO_API_CLIENT_SECRET,
+    scopes: [],
+    state: ''
+});
 
 setInterval(
     async () => {
@@ -29,12 +37,12 @@ setInterval(
         count += 1;
 
         try {
-            debug('count:', count);
             await sskts.service.task.executeByName(
-                sskts.factory.taskName.UseMvtk
+                sskts.factory.taskName.ReturnPointAward
             )({
                 taskRepo: taskRepo,
-                connection: sskts.mongoose.connection
+                connection: mongoose.connection,
+                pecorinoAuthClient: authClient
             });
         } catch (error) {
             // tslint:disable-next-line:no-console