+++ src/jobs/triggered/createImportScreeningEventsTask/run.ts
@@ -4,6 +4,7 @@
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
 import * as moment from 'moment';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../mongooseConnectionOptions';
 
@@ -20,14 +21,14 @@ const LENGTH_IMPORT_SCREENING_EVENTS_IN_WEEKS = (process.env.LENGTH_IMPORT_SCREE
 
 async function main() {
     debug('connecting mongodb...');
-    await sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+    await mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
-    const placeRepo = new sskts.repository.Place(sskts.mongoose.connection);
-    const organizationRepo = new sskts.repository.Organization(sskts.mongoose.connection);
-    const taskRepo = new sskts.repository.Task(sskts.mongoose.connection);
+    const placeRepo = new sskts.repository.Place(mongoose.connection);
+    const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+    const taskRepo = new sskts.repository.Task(mongoose.connection);
 
     // 全劇場組織を取得
-    const movieTheaterOrganizations = await organizationRepo.searchMovieTheaters({});
+    const sellers = await sellerRepo.search({});
     const movieTheaters = await placeRepo.searchMovieTheaters({});
     const importFrom = moment()
         .toDate();
@@ -38,22 +39,34 @@ async function main() {
 
     await Promise.all(movieTheaters.map(async (movieTheater) => {
         try {
-            const movieTheaterOrganization = movieTheaterOrganizations.find((m) => m.location.branchCode === movieTheater.branchCode);
-            if (movieTheaterOrganization !== undefined) {
+            const branchCode = movieTheater.branchCode;
+            const seller = sellers.find((m) => {
+                return m.location !== undefined
+                    && m.location.branchCode !== undefined
+                    && m.location.branchCode === branchCode;
+            });
+
+            if (seller !== undefined) {
+                let xmlEndPoint: any;
+                if (Array.isArray((<any>seller).additionalProperty)) {
+                    const xmlEndPointProperty = (<any>seller).additionalProperty.find(((p: any) => {
+                        return p.name === 'xmlEndPoint';
+                    }));
+                    xmlEndPoint = (xmlEndPointProperty !== undefined) ? JSON.parse(xmlEndPointProperty.value) : undefined;
+                }
+
                 const taskAttributes: sskts.factory.task.IAttributes<sskts.factory.taskName.ImportScreeningEvents> = {
                     name: sskts.factory.taskName.ImportScreeningEvents,
                     status: sskts.factory.taskStatus.Ready,
                     runsAt: runsAt,
                     remainingNumberOfTries: 1,
-                    // tslint:disable-next-line:no-null-keyword
-                    lastTriedAt: null,
                     numberOfTried: 0,
                     executionResults: [],
                     data: {
-                        locationBranchCode: movieTheaterOrganization.location.branchCode,
+                        locationBranchCode: branchCode,
                         importFrom: importFrom,
                         importThrough: importThrough,
-                        xmlEndPoint: movieTheaterOrganization.xmlEndPoint
+                        xmlEndPoint: xmlEndPoint
                     }
                 };
                 await taskRepo.save(taskAttributes);
@@ -68,7 +81,7 @@ async function main() {
     await new Promise((resolve) => {
         setTimeout(
             async () => {
-                await sskts.mongoose.disconnect();
+                await mongoose.disconnect();
                 resolve();
             },
             // tslint:disable-next-line:no-magic-numbers