+++ src/jobs/triggered/importMovieTheaters/run.ts
@@ -3,6 +3,7 @@
  */
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../mongooseConnectionOptions';
 
@@ -10,31 +11,32 @@ const debug = createDebug('sskts-jobs:jobs');
 
 async function main() {
     debug('connecting mongodb...');
-    await sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+    await mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
-    const organizationRepo = new sskts.repository.Organization(sskts.mongoose.connection);
-    const placeRepo = new sskts.repository.Place(sskts.mongoose.connection);
+    const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+    const placeRepo = new sskts.repository.Place(mongoose.connection);
 
     // 全劇場組織を取得
-    const movieTheaters = await organizationRepo.searchMovieTheaters({});
+    const sellers = await sellerRepo.search({});
 
-    for (const movieTheater of movieTheaters) {
-        const branchCode = movieTheater.location.branchCode;
-
-        try {
-            debug('importing movieTheater...', branchCode);
-            await sskts.service.masterSync.importMovieTheater(branchCode)({
-                organization: organizationRepo,
-                place: placeRepo
-            });
-            debug('movieTheater imported', branchCode);
-        } catch (error) {
-            // tslint:disable-next-line:no-console
-            console.error(error);
+    for (const seller of sellers) {
+        if (seller.location !== undefined && seller.location.branchCode !== undefined) {
+            try {
+                const branchCode = seller.location.branchCode;
+                debug('importing movieTheater...', branchCode);
+                await sskts.service.masterSync.importMovieTheater(branchCode)({
+                    seller: sellerRepo,
+                    place: placeRepo
+                });
+                debug('movieTheater imported', branchCode);
+            } catch (error) {
+                // tslint:disable-next-line:no-console
+                console.error(error);
+            }
         }
     }
 
-    await sskts.mongoose.disconnect();
+    await mongoose.disconnect();
 }
 
 main()