+++ src/jobs/triggered/importMovies/run.ts
@@ -3,6 +3,7 @@
  */
 import * as sskts from '@motionpicture/sskts-domain';
 import * as createDebug from 'debug';
+import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../../mongooseConnectionOptions';
 
@@ -10,29 +11,30 @@ const debug = createDebug('sskts-jobs:jobs');
 
 export async function main() {
     debug('connecting mongodb...');
-    await sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
+    await mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
-    const creativeWorkRepository = new sskts.repository.CreativeWork(sskts.mongoose.connection);
-    const organizationRepository = new sskts.repository.Organization(sskts.mongoose.connection);
+    const creativeWorkRepository = new sskts.repository.CreativeWork(mongoose.connection);
+    const sellerRepo = new sskts.repository.Seller(mongoose.connection);
 
     // 全劇場組織を取得
-    const movieTheaters = await organizationRepository.searchMovieTheaters({});
+    const sellers = await sellerRepo.search({});
 
     // 劇場ごとに映画作品をインポート
-    for (const movieTheater of movieTheaters) {
-        const branchCode = movieTheater.location.branchCode;
-
-        try {
-            debug('importing movies...', branchCode);
-            await sskts.service.masterSync.importMovies(branchCode)({ creativeWork: creativeWorkRepository });
-            debug('movies imported', branchCode);
-        } catch (error) {
-            // tslint:disable-next-line:no-console
-            console.error(error);
+    for (const seller of sellers) {
+        if (seller.location !== undefined && seller.location.branchCode !== undefined) {
+            try {
+                const branchCode = seller.location.branchCode;
+                debug('importing movies...', branchCode);
+                await sskts.service.masterSync.importMovies(branchCode)({ creativeWork: creativeWorkRepository });
+                debug('movies imported', branchCode);
+            } catch (error) {
+                // tslint:disable-next-line:no-console
+                console.error(error);
+            }
         }
     }
 
-    await sskts.mongoose.disconnect();
+    await mongoose.disconnect();
 }
 
 main()