+++ src/mongooseConnectionOptions.ts
@@ -2,10 +2,11 @@
  * mongoose接続オプション
  * @see http://mongoosejs.com/docs/api.html#index_Mongoose-connect
  */
-import * as sskts from '@motionpicture/sskts-domain';
-const mongooseConnectionOptions: sskts.mongoose.ConnectionOptions = {
+import * as mongoose from 'mongoose';
+
+const mongooseConnectionOptions: mongoose.ConnectionOptions = {
     autoReconnect: true,
-    keepAlive: 120000,
+    keepAlive: true,
     connectTimeoutMS: 30000,
     socketTimeoutMS: 0,
     reconnectTries: 30,