+++ CHANGELOG.md
@@ -18,6 +18,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v8.0.0 - 2019-02-05
+
+### Changed
+
+- 組織のxmlEndPointをadditionalPropertyから読み出すように変更
+- install @motionpicture/sskts-domain@28.0.0
+- タスク名をCinerinoに統一
+- クレジットカード決済インターフェースをCinerinoに統一
+- クレジットカード返金インターフェースをCinerinoに統一
+- 注文タスクインターフェースをCinerinoに統一
+- 注文配送タスクインターフェースをCinerinoに統一
+- 返品タスクインターフェースをCinerinoに統一
+- イベントインターフェースをCinerinoに統一
+- 予約インターフェースをCinerinoに統一
+- 注文インターフェースをCinerinoに統一
+
 ## v7.0.0 - 2019-01-28
 
 ### Changed